solve(4): median of two sorted arrays
uses binary search on a single array, with the insight of what makes a number a
medium in these sorted lists. Note that we only search one list single the
index for the other list and be automatically determined. It's also important
to handle edge cases (setting to inf or -inf) or practical purposes

Tags: Binary-Search