solve(739): daily temperatures - medium
use monotonic stack that calculates the temperature at every pop.
note that stack is sort of a delayed processing 'queue',
we wait for the right conditions then we calculate the values.

this question can also be solved with direct address map/hash map
since the range of possible values if small (only 71)

Tags: monotonic-stack hash-table direct-address-table