solve(1071): gcd of strings - easy
the elegant solution is quite advanced, which requires proofs to establish.
it utilizes the existence of gcd string by the structure of str1 and
str2. then argues that the gcd of the lengths of the string is the
length of the gcd string. Reducing the question to just solving a gcd
problem.

Tags: Math