solve(502): IPO max profit with capital limit - hard
an intuitive greedy answer, but implementation isn't that straight
forward. When I think of using heap, the first thing I do is to populate
it with all the data and at most store a pointer in it that ref to some
other data structure. REMEMBER that heap can also be used to track the
current max/min. try not to use binary search in situation where I need
to keep track of more than two sets of data structure that requires
max/min or in some order.

all in all, not a trivial task to achieve a complexity of O(n logn)

Tags: Greedy Sorting Heap