solve(652): find structural duplicate subtree - medium
pretty hard in fact, the key is to find a good hashable representation
of a tree. Now we can either use the string representation, which
contains information on the entire tree, or we can use triplet & ID
representation. Then all we need to do is to count the occurrences of
these IDs and push the subtree accordingly

Tags: Hash-Table Binary-Tree DFS