solve(109): convert sorted list to balanced BST - medium
interesting question with several possible solutions. using Divide and
Conquer with list/vec is the naive way of doing this as it creates a
bunch of copies by slicing. I made the mistake of modifying the
original LinkedList while traversing, and it mutated the input. It's
acceptable in python, but not in rust. so i have reimplemented the
solution is use two pointers as position to start and stop traversing
the LinkedList.

Tags: BST Linked-List Divide-and-Conquer