solve(958): is complete binary tree - medium
Could use BFS with observation of no null node after we encounter the
first null node. Use DFS with observation that index of left child is (2 * i +
1) and right child is (2 * i + 2), while no node's index can be >= to n
within a complete tree, since they are tightly pact. (need to run DFS
first to get value of n)

Also, could solve it my way using information of the completeness,
fullness and height of child to determine the same of root. Albeit there
were a couple complex cases that needs to considered, thus the solution
is not as elegant, but it is a one pass DFS algorithm, instead of the
previous 2 pass mentioned above.

Tags: DFS BFS